refactor(settings): Updated settings page structure
