+++ src/renderer/src/pages/settings/MiniappSettings/MiniAppIconsManager.tsx
